Extract Enum into its own module to avoid circular dependencies
